requirements.txt
# python 3.12.10
certifi-2025.8.3
charset_normalizer-3.4.2
geopandas-1.1.1
idna-3.10
numpy-2.3.2
osmium-4.0.2
packaging-25.0
pandas-2.3.1
pyogrio-0.11.1
pyproj-3.7.1
python-dateutil-2.9.0.post0
pytz-2025.2
requests-2.32.4
shapely-2.1.1
six-1.17.0
tzdata-2025.2
urllib3-2.5.0
Flask>=3.0
folium>=0.18.0
